03-product-service/requirements.txt
Flask==2.3.2
psutil==5.9.5
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-flask==0.41b0

